use set seed for random operations

+++ include/types.hpp
@@ -2,7 +2,7 @@
 
 #include <Eigen/Dense>
 
-using RealType = double;
+using RealType = float;
 
 using DynamicMatrix = Eigen::Matrix<RealType, Eigen::Dynamic, Eigen::Dynamic>;
 using DynamicVector = Eigen::Matrix<RealType, Eigen::Dynamic, 1>;
